backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
aioredis==2.0.1
redis==5.2.1
apscheduler==3.10.4
python-socketio==5.12.0
httpx==0.28.1
twilio==9.4.0
sendgrid==6.11.0
timezonefinder==6.5.7
geopy==2.4.1
aiofiles==24.1.0
email-validator==2.2.0
pytz==2025.2
